Move all test timer tests to a spearate test module
