+++ __tests__/componentTests/YouTubePlayer.test.tsx
@@ -1,7 +1,11 @@
 import { render, screen } from '@testing-library/react'
 import userEvent from '@testing-library/user-event';
 import YouTubePlayer from '../../components/YouTubePlayer';
-import { act } from 'react-dom/test-utils';
+
+// Named mocks to test player functions being called
+const muteMock = jest.mock;
+const unMuteMock = jest.fn;
+
 
 // Provide channel data and other UI states via this mock of the channel search API call
 jest.mock('../../hooks/useYouTubeIframe', () => ({
@@ -10,8 +14,8 @@ jest.mock('../../hooks/useYouTubeIframe', () => ({
     player: {
       getCurrentTime: () => 0,
       isMuted: jest.fn,
-      mute: jest.fn,
-      unMute: jest.fn,
+      mute: muteMock,
+      unMute: unMuteMock,
     }
   })
 }));
@@ -124,26 +128,23 @@ describe('YouTube player control toggles', () => {
     const customControls = screen.getByTestId('customControls');
     expect(customControls).not.toHaveClass('controlsHide');
   });
+});
 
-  it('Hides custom controls 3 seconds after user activity is first registered', async () => {
+describe('YouTube player keyboard shortcuts', () => {
+  it('Mutes the video on "m" key press, when in an unmuted state', async () => {
     render(<YouTubePlayer videoId='1234' />)
     const hideYTBtn = screen.getByRole('button', { name: /hide YT controls/i });
+    const wrapper = screen.getByTestId('wrapper');
 
-    // First enable custom controls, then hover the relevant div to trigger user activity/controls to show
+    // First enable custom controls, then focus the wrapper to ensure the keypress is captured correctly
     await userEvent.click(hideYTBtn);
-    const overlay = screen.getByTestId('overlay');
-    await userEvent.hover(overlay);
+    wrapper.focus();
+    await userEvent.keyboard('m');
 
-    // Wait for controls to fade. Act is called here because this line does note directly use React Testing Library, and is the line that involves several DOM elements changing/re-rendering. React recommends act() in these cases
-    await act(async () => {
-      await new Promise(res => setTimeout(res, 3500));
-    })
-
-    const customControls = screen.getByTestId('customControls');
-    expect(customControls).toHaveClass('controlsHide');
+    expect(muteMock).toBeCalled();
   });
 
-  it('Shows controls on press of mute keyboard shortcut', async () => {
+  it('Unutes the video on "m" key press, when in a muted state', async () => {
     render(<YouTubePlayer videoId='1234' />)
     const hideYTBtn = screen.getByRole('button', { name: /hide YT controls/i });
     const wrapper = screen.getByTestId('wrapper');
@@ -156,56 +157,5 @@ describe('YouTube player control toggles', () => {
     const customControls = screen.getByTestId('customControls');
     expect(customControls).not.toHaveClass('controlsHide');
   });
-
-  it('Cancels original fade out timer when userr is active within original fade out timer', async () => {
-    render(<YouTubePlayer videoId='1234' />)
-    const hideYTBtn = screen.getByRole('button', { name: /hide YT controls/i });
-
-    // First enable custom controls, then hover the relevant div to trigger user activity/controls to show
-    await userEvent.click(hideYTBtn);
-    const overlay = screen.getByTestId('overlay');
-    await userEvent.hover(overlay);
-
-    // Wait only 0.5 seconds before hovering again to re-trigger controls
-    await act(async () => {
-      await new Promise(res => setTimeout(res, 1500));
-    });
-
-    await userEvent.hover(overlay);
-
-    // Check during the period that the original timer would have ended
-    await act(async () => {
-      await new Promise(res => setTimeout(res, 2000));
-    });
-
-    const customControls = screen.getByTestId('customControls');
-    expect(customControls).not.toHaveClass('controlsHide');
-  });
-
-  it('Resets fade out time when user is active during original fade out time', async () => {
-    render(<YouTubePlayer videoId='1234' />)
-    const hideYTBtn = screen.getByRole('button', { name: /hide YT controls/i });
-
-    // First enable custom controls, then hover the relevant div to trigger user activity/controls to show
-    await userEvent.click(hideYTBtn);
-    const overlay = screen.getByTestId('overlay');
-    await userEvent.hover(overlay);
-
-    // Wait only 0.5 seconds before hovering again to re-trigger controls
-    await act(async () => {
-      await new Promise(res => setTimeout(res, 500));
-    });
-
-    await userEvent.hover(overlay);
-
-    // Wait a further 3 seconds for controls to disappear
-    await act(async () => {
-      await new Promise(res => setTimeout(res, 3500));
-    });
-
-    const customControls = screen.getByTestId('customControls');
-    expect(customControls).toHaveClass('controlsHide');
-  });
 });
-
